minor update: static
When a random number generator is declared with local scope (i.e. within functions) or is not declared static there is a risk that the same random numbers will be reused over and over. It would be better to define this globally and static.

static    default_random_engine gen;

To see the effect first hand, change the number of particles to 1. You will see the estimate (the particle position) drift off continuously in a constant direction as the same errors are added repeatedly: they will not average out to zero.